[Refactor] main-today, main-week 폴더명을 home-today, home-week로 변경
