+++ src/pages/main/home-today/DailyInfo.tsx
