+++ src/pages/main/home-today/DailyIntake.tsx
