+++ src/pages/main/home-week/WeeklyInfo.tsx
